backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1
rq==2.0.0
pydantic-settings==2.6.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
boto3==1.35.0
httpx==0.27.0
